pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "irodori-character-pipeline"
version = "0.1.0"
description = "Shared tools for manuscript-driven character profile extraction and Irodori VoiceDesign sample generation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "gradio_client>=1.0.0",
  "httpx>=0.27.0",
]

[tool.setuptools]
py-modules = []
